backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
celery[redis]==5.4.0
redis==5.0.0
openai==1.50.0
langgraph
langchain-openai
langchain-core
playwright==1.47.0
httpx==0.27.0
beautifulsoup4==4.12.0
pydantic==2.9.0
python-dotenv==1.0.0
aiosqlite==0.20.0
Pillow==10.4.0